# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTrack API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "The AquaTrack API allows users to log their daily water intake, edit or delete records, and track their progress towards a daily hydration goal. Users can view their current day's water consumption and analyze their intake over the month."
  },
  "servers": [
    {
      "url": "https://aqua-track-app.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "description": "Creates a new user account and returns email and authentication token.",
        "security": [],
        "requestBody": {
          "description": "User registration data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "description": "JTestemail@gmail.com",
                    "type": "string",
                    "example": "testemail@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 50,
                    "pattern": "^(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])",
                    "example": "Example@1",
                    "description": "Password must be between 6 and 50 characters in length.\nIt must contain at least one uppercase letter, one number, and one special character from the following: !@#$%^&*.\n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered a user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully registered a new user!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/auth/verifycate/{token}": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify email",
        "operationId": "verifyEmail",
        "description": "Verifies the user's email address using the provided token.",
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2N2Q0YWY4MWM4YWVhZTY5YTFiZWJkNjIiLCJlbWFpbCI6Imtyb2xpYXRrb29AZ21haWwuY29tIiwiaWF0IjoxNzQxOTkxODE4LCJleHAiOjE3NDE5OTI3MTh9.inRLWLNU0tRsNaaxbfG3TeietOXe8f3HrGrk6ybagOA"
            },
            "required": true,
            "description": "Verification token received in the email."
          }
        ],
        "responses": {
          "200": {
            "description": "Email successfully verified!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "status"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email successfully verified!"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login to the account",
        "operationId": "loginUser",
        "description": "Logs in a user and returns an authentication token.",
        "security": [],
        "requestBody": {
          "description": "User login data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "description": "User's email",
                    "type": "string",
                    "example": "JTestemail@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 50,
                    "pattern": "^(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])",
                    "example": "Example@1",
                    "description": "Password must be between 6 and 50 characters in length.\nIt must contain at least one uppercase letter, one number, and one special character from the following: !@#$%^&*.\n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in a user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "bseDlf/Chq0dMG6c6o194qTqA6KaFeMMt6lJZOhp"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout from the account",
        "operationId": "logoutUser",
        "description": "User logout, removes the sessionId and refreshToken cookies, and also deletes the session record from the database.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "operationId": "resetPassword",
        "description": "Resets password with the token from email.\n",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2N2QzODUxZGRlMTM2MWE5OTZjZmIwZmUiLCJlbWFpbCI6Imtyb2xpYXRrb29AZ21haWwuY29tIiwiaWF0IjoxNzQxOTY0OTAxLCJleHAiOjE3NDE5NjU4MDF9.m8_RfZmHECXGjw__gCtEVKhJ-vYhJiOp1NdyywXCKBM",
                    "description": "Token for reseting password"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 50,
                    "pattern": "^(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])",
                    "example": "Example@1",
                    "description": "Password must be between 6 and 50 characters in length.\nIt must contain at least one uppercase letter, one number, and one special character from the following: !@#$%^&*.\n"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password was successfully reset!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Reset password email was successfully sent!"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "object"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "jwt expired"
                    },
                    "data": {
                      "oneOf": [
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/user/update/{id}": {
      "patch": {
        "summary": "Update user details",
        "description": "Updates the user profile, allowing changes to fields such as name, email, gender, weight, and daily water intake goal.",
        "tags": [
          "Users"
        ],
        "operationId": "updateUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the user to update",
            "schema": {
              "type": "string",
              "example": "67cc5bb27a2222d660c179f9"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 12,
                    "example": "JohnDoe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "johndoe@example.com"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "woman",
                      "man"
                    ],
                    "example": "woman"
                  },
                  "weight": {
                    "type": "number",
                    "minimum": 0,
                    "example": 75
                  },
                  "dailyNorm": {
                    "type": "number",
                    "minimum": 500,
                    "maximum": 15000,
                    "example": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User successfully updated!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "_id": {
            "description": "User's ID",
            "type": "string",
            "example": "67cc5bb27a2222d660c179f9"
          },
          "name": {
            "description": "User's name",
            "type": "string",
            "minLength": 3,
            "maxLength": 12,
            "example": "JohnDoe"
          },
          "email": {
            "description": "User's email",
            "type": "string",
            "format": "email",
            "example": "JTestemail@gmail.com"
          },
          "password": {
            "description": "User's password",
            "type": "string",
            "example": "password12345!"
          },
          "gender": {
            "description": "User's gender",
            "type": "string",
            "enum": [
              "woman",
              "man"
            ],
            "example": "woman"
          },
          "weight": {
            "description": "User's weight",
            "type": "number",
            "minimum": 0,
            "example": 60
          },
          "dailySportTime": {
            "description": "User's daily sport time in minutes",
            "type": "integer",
            "minimum": 0,
            "example": 1
          },
          "dailyNorm": {
            "description": "User's daily water intake norm in ml",
            "type": "number",
            "minimum": 500,
            "maximum": 15000,
            "example": 2000
          },
          "avatar": {
            "description": "User's avatar URL",
            "type": "string",
            "example": "https://res.cloudinary.com/diee7l1on/image/upload/v1741187439/default-avatar-profile-icon-social-media-user-image-gray-avatar-icon-blank-profile-silhouette-vector-illustration_561158-3467_bvqe8k.avif"
          },
          "isVerified": {
            "description": "User's verification status",
            "type": "boolean",
            "example": false
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Missing or invalid data"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Missing or invalid request parameters."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 403
                },
                "message": {
                  "type": "string",
                  "example": "Email is not verified."
                },
                "data": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "NotFoundError",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The requested resource does not exist."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "409": {
        "description": "Conflict error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "ConflictError"
                },
                "data": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email in use"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "data": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "type": "object"
                    }
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Something went wrong"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}